requirements.txt
pymongo.py
discord.py
os.py
json.py
asyncio.py

main.py
torn_api.py
mongo_db.py
command_handler.py
bot_actions.py